Spec behavior for collapsed selections
Which is . . . nothing.  Invoking bold (for instance) when the selection
doesn't contain anything just does nothing.  I expected it to insert an
empty <b></b> with the cursor inside so that if you start typing it's
bold.  But apparently, that doesn't happen.  So do editors have to fake
up that effect themselves, or am I missing something?  Puzzling.