Don't allow merging from/to table cells at all
I changed my mind, partly based on the recommendation of the "behavior
when typing in contentEditable elements" document.  I was ambivalent to
begin with, but on consideration, the new behavior is more reasonable.

I also refactored things in a few ways that should be unimportant,
although I'm pretty sure I fixed at least one latent bug.