Double-test everything for conformance tests
Except multitest.  The CSS styling flag is relevant to things like
delete and insertText and whatnot as well as commands that actually
refer to it directly.  Browsers might also care about it in violation of
spec, like for hiliteColor.

Needless to say, I didn't review the whole diff to data.js.  I only
looked at a few places.